+++ src/main/java/com/bleudev/nine_lifes/client/compat/modmenu/NineLifesConfig.java
@@ -7,4 +7,10 @@ public class NineLifesConfig extends MidnightConfig {
     public static boolean join_message_enabled = true;
     @Entry
     public static boolean heartbeat_enabled = true;
+
+    public enum HeartPosition {
+        BOTTOM_LEFT, BOTTOM_CENTER, BOTTOM_RIGHT, TOP_LEFT, TOP_CENTER, TOP_RIGHT
+    }
+    @Entry
+    public static HeartPosition heart_position = HeartPosition.BOTTOM_CENTER;
 }