+++ src/main/java/com/bleudev/nine_lifes/custom/consume/AmethysmConsumeEffect.java
@@ -15,7 +15,6 @@ import net.minecraft.item.consume.ConsumeEffect;
 import net.minecraft.network.RegistryByteBuf;
 import net.minecraft.network.codec.PacketCodec;
 import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
 import net.minecraft.world.World;
@@ -28,7 +27,6 @@ public record AmethysmConsumeEffect() implements ConsumeEffect {
     public static final MapCodec<AmethysmConsumeEffect> CODEC = MapCodec.unit(INSTANCE);
     public static final PacketCodec<RegistryByteBuf, AmethysmConsumeEffect> PACKET_CODEC = PacketCodec.unit(INSTANCE);
 
-
     @Override
     public Type<? extends ConsumeEffect> getType() {
         return AMETHYSM_CONSUME_EFFECT_TYPE;
@@ -40,14 +38,12 @@ public record AmethysmConsumeEffect() implements ConsumeEffect {
             .anyMatch(entry -> entry.getKey()
                 .map(key -> key.equals(CustomEnchantments.CHARGE))
                 .orElse(false))) {
-            if (world instanceof ServerWorld serverWorld) {
-                if (entity instanceof ServerPlayerEntity player) {
-                    LivesUtils.setLives(player, o -> o + 2);
-                    ServerPlayNetworking.send(player, new StartWhitenessScreenEvent(TICKS_PER_SECOND));
-                }
-
-                ((LivingEntityCustomInterface) entity).nine_lifes$setDamageTicks(TICKS_PER_SECOND);
+            if (entity instanceof ServerPlayerEntity player) {
+                LivesUtils.setLives(player, o -> o + 2);
+                ServerPlayNetworking.send(player, new StartWhitenessScreenEvent(TICKS_PER_SECOND));
             }
+
+            ((LivingEntityCustomInterface) entity).nine_lifes$setDamageTicks(TICKS_PER_SECOND);
             return true;
         }